Optimize code structure.
