--- shardingsphere-features/shardingsphere-shadow/shardingsphere-shadow-core/src/test/java/org/apache/shardingsphere/shadow/algorithm/shadow/ShadowAlgorithmTypeTest.java
@@ -1,40 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one or more
- * contributor license agreements.  See the NOTICE file distributed with
- * this work for additional information regarding copyright ownership.
- * The ASF licenses this file to You under the Apache License, Version 2.0
- * (the "License"); you may not use this file except in compliance with
- * the License.  You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.apache.shardingsphere.shadow.algorithm.shadow;
-
-import org.junit.Test;
-
-import java.util.Optional;
-
-import static org.hamcrest.CoreMatchers.is;
-import static org.junit.Assert.assertThat;
-
-public final class ShadowAlgorithmTypeTest {
-    
-    @Test
-    public void assertNewInstance() {
-        Optional<ShadowAlgorithmType> columnRegexMatch = ShadowAlgorithmType.newInstance("COLUMN_REGEX_MATCH");
-        assertThat(columnRegexMatch.isPresent(), is(true));
-        assertThat(columnRegexMatch.get(), is(ShadowAlgorithmType.COLUMN_REGEX_MATCH));
-        Optional<ShadowAlgorithmType> simpleNote = ShadowAlgorithmType.newInstance("SIMPLE_NOTE");
-        assertThat(simpleNote.isPresent(), is(true));
-        assertThat(simpleNote.get(), is(ShadowAlgorithmType.SIMPLE_NOTE));
-        Optional<ShadowAlgorithmType> autoType = ShadowAlgorithmType.newInstance("AUTO_TYPE");
-        assertThat(autoType.isPresent(), is(false));
-    }
-}
